Ticker Model updates
Updated the Ticker model class.
